+++ firmware/application/stm32g474/modules/chassis/odometry.h
@@ -6,12 +6,12 @@
 typedef struct {
   int32_t left_delta;
   int32_t right_delta;
-  int32_t left_total;
-  int32_t right_total;
+  int64_t left_total;
+  int64_t right_total;
   int32_t forward_count_sum_delta;
   int32_t turn_count_difference_delta;
-  int32_t forward_count_sum_total;
-  int32_t turn_count_difference_total;
+  int64_t forward_count_sum_total;
+  int64_t turn_count_difference_total;
 } OdometrySnapshot;
 
 void Odometry_Init(void);
